AI attendance/backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.1
pydantic[email]>=2.9.0
motor>=3.6.0
beanie>=1.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
opencv-python-headless>=4.10.0.84
deepface>=0.0.93
numpy>=1.26.4
requests>=2.32.3
